Removes unneeded include.
The header <limits> is not needed here, since <type_traits> does the job with is_integral and is_floating_point. For those without access to a C++11 compiler, <limits> can be used in combination with Boost enable_if, in which case they need to check that std::numeric_limits<T>::is_integral be either true or false.